Move all DTO VO and entities to model package.
