+++ src/main/java/io/github/jerrychin/testbackendjava/controller/UserController.java
@@ -1,8 +1,8 @@
 package io.github.jerrychin.testbackendjava.controller;
 
 
-import io.github.jerrychin.testbackendjava.dto.UserDTO;
-import io.github.jerrychin.testbackendjava.dto.UserVO;
+import io.github.jerrychin.testbackendjava.model.dto.UserDTO;
+import io.github.jerrychin.testbackendjava.model.vo.UserVO;
 import io.github.jerrychin.testbackendjava.service.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
