+++ src/main/java/io/github/jerrychin/testbackendjava/model/dto/AccessTokenDTO.java
@@ -1,4 +1,4 @@
-package io.github.jerrychin.testbackendjava.dto;
+package io.github.jerrychin.testbackendjava.model.dto;
 
 public class AccessTokenDTO {
