+++ src/main/java/io/github/jerrychin/testbackendjava/model/vo/PeopleVO.java
@@ -1,4 +1,4 @@
-package io.github.jerrychin.testbackendjava.dto;
+package io.github.jerrychin.testbackendjava.model.vo;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
