+++ src/main/java/io/github/jerrychin/testbackendjava/model/vo/UserVO.java
@@ -1,34 +1,24 @@
-package io.github.jerrychin.testbackendjava.entity;
+package io.github.jerrychin.testbackendjava.model.vo;
 
-import org.hibernate.annotations.GenericGenerator;
+import com.fasterxml.jackson.annotation.JsonFormat;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-@Entity
-@Table(name = "user")
-public class User {
+public class UserVO {
 
-    @Id
-    @GenericGenerator(name = "user_id", strategy = "io.github.jerrychin.testbackendjava.generator.UserIdGenerator")
-    @GeneratedValue(generator = "user_id")
     private String id;
 
-    private LocalDateTime createdAt;
-
     /**
-     * unique user name.
+     * name.
      */
     private String name;
 
     /**
      * date of birth.
      */
+    @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate dob;
 
     /**
@@ -38,14 +28,12 @@ public class User {
 
     private String description;
 
-    /**
-     * associated account id.
-     */
-    private Long accountId;
-
     private BigDecimal longitude;
 
     private BigDecimal latitude;
+
+    @JsonFormat(pattern = "yyyy-MM-dd hh:mm:ss")
+    private LocalDateTime createdAt;
 
     public String getId() {
         return id;
@@ -93,14 +81,6 @@ public class User {
 
     public void setDescription(String description) {
         this.description = description;
-    }
-
-    public Long getAccountId() {
-        return accountId;
-    }
-
-    public void setAccountId(Long accountId) {
-        this.accountId = accountId;
     }
 
     public BigDecimal getLongitude() {
