+++ src/main/java/io/github/jerrychin/testbackendjava/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package io.github.jerrychin.testbackendjava.repository;
 
-import io.github.jerrychin.testbackendjava.entity.User;
+import io.github.jerrychin.testbackendjava.model.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
