+++ src/test/java/io/github/jerrychin/testbackendjava/service/AccountServiceTest.java
@@ -1,8 +1,8 @@
 package io.github.jerrychin.testbackendjava.service;
 
 import io.github.jerrychin.testbackendjava.SampleBaseTestCase;
-import io.github.jerrychin.testbackendjava.dto.AccountDTO;
-import io.github.jerrychin.testbackendjava.entity.Account;
+import io.github.jerrychin.testbackendjava.model.dto.AccountDTO;
+import io.github.jerrychin.testbackendjava.model.entity.Account;
 import io.github.jerrychin.testbackendjava.repository.AccountRepository;
 import org.junit.jupiter.api.Test;
 import org.mockito.*;
